requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
requests
psycopg2-binary
redis
groq
biopython
python-multipart
pillow
pyzbar
gTTS
opencv-python
